Fixed capacity issue
Matlab is shit. By default the slx model doesn't inherit the cell characteristics, so they have to be altered manually. To do this I edited the .ssc files in +Battery to correct the Cell Capacity. I also updated the SoC, V0 and R0 to be accurate to the VTC6 cell we use.